Fix, simplify and document IndexedEmbedded schema filtering
The fixes:

* We used to allow an IndexedEmbedded inside another IndexedEmbedded with
maxDepth = 1, which is wrong.
* IndexSchemaNestingContextImpl used to pass the wrong field names to the
filter, due to it prepending the IndexedEmbedded prefix where it should
not.
* We used to filter out implicitly included paths from a nested filter
even when they were explicitly included by the parent filter as soon as
the parent filter had a maxDepth of 0, which is wrong.

The simplification and documentation are an attempt to make the whole
schema filtering code more understandable, because those bugs were quite
hard to spot and understand, and I sincerly wish future me (or whoever
has to deal with this) an easier time debugging this.