Split PojoModelElement into two interfaces: PojoModelType and PojoModelProperty
This allows to add methods specific to a type model or to a property
model. As an example, I added a getName() method to the
PojoModelProperty, which could be used to set the default name of a
field.

Also, differentiate the TypeBridge and PropertyBridge interfaces by
expecting a PojoModelType in the former's bind method and a
PojoModelProperty in the latter's.