Make sure to trim down the processor tree according to the schema filtering
Schema filtering occurs when an IndexedEmbedded has non-default
parameters, for instance a max depth or explicitly included paths.

It may prevent bridges to add certain fields to the schema, and
depending on the filters may even prevent some bridges to add *any*
field to the schema. When that happens, the bridges are effectively
useless, since they will never add anything to the index, and thus we
don't need to keep them.

Ignoring those bridges may result in some Pojo processors to become
empty (they don't use the value they extract from the Pojo), in which
case we can also trim those processors to reduce the number of
accessed properties during indexing.