HSEARCH-3123 Elasticsearch - Move backend-scoped resources for search to a dedicated object
We don't want to pass lots of parameters everywhere in the classes
implementing search, but we don't want these classes to have a
dependency to a common class (the backend) which could reference tons of
other classes we don't need for search.

Thus, the solution is to introduce a specific class that declares
all that's needed for search, and let the backend instantiate it.