HSEARCH-3199 Avoid redundant filtering nodes in reindexing resolvers
This is a bit hard to test, since the only effect of this patch is to
avoid a few calls to PojoDirtinessState#isAnyDirty when we know it will
return true. There should be no effect whatsoever on the indexing
process otherwise: the same properties will be accessed, in the same
order, producing the same document, etc.

In short, this is just a performance improvement.