HSEARCH-3208 Use an integer percentage instead of a double ratio for relative minimumShouldMatch constraints
This is important in particular because of rounding errors inherent to
floating-point arithmetic, which is especially visible when we need to
round the result. Note this is true even for small numbers.

For example, in Java, 0.58 * 50 = 28.999999999999996 which rounds down
to 28, whereas 58 * 50 / 100 = 29.

For the record, the following command can be run to find similar
rounding errors by brute force. Java 9+ required to use jshell.

jshell <<EOF
for ( int i = 0; i < 1000; ++i ) {
        for ( int j = 0; j <= 100; ++j ) {
                int asDouble = (int) ( ( (double) j / 100.0 ) * i );
                int asInteger = j * i / 100;
                if ( asDouble != asInteger ) {
                        System.out.println( "Different: " + j + "% of " + i + " = " + asDouble + " as double or " + asInteger + " as integer" );
                }
        }
}
EOF