HSEARCH-3199 Use simple strings to represent dirty paths in automatic reindexing
Because it is simpler, and because we certainly do not want to build
PojoModelPath instances at runtime.

Ideally, at some point in the future, we will want to use more advanced
filters that are closer to the original entity metadata. For example in
the ORM integration, dirty properties are stored as an array of
integers. It would be better if we could just pass this array, and store
dirty properties the same way.

But that would require to allow mapper implementations such as ORM to
change the worker APIs, and that would require a lot more change, so
let's keep it for later.

I tried to keep things ready for later changes, in particular by
encapsulating the concept of filtering in a dedicated interface, by
making the mapper implementations (ORM, JavaBean) provide the filter
themselves, and by making the type of the "dirtiness state" generic,
so that the automatic reindexing code will not require any further
change. The PojoMapper/PojoSearchManager/PojoChangesetWorker code
will require changes, though.