HSEARCH-2546 Use the programmatic mapping in the annotation mapping
More precisely, change the annotation mapping so that it only parses
annotations and forwards the data to the programmatic mapping classes,
which will store the information and contribute it later (potentially
multiple times).

The main reason for this change was originally to avoid re-implementing
exception catching and contextual failure collection in both the
annotation mapping and the programmatic mapping: we will be able to only
implement it in the programmatic mapping, and the annotation mapping
will benefit from it automatically.

But there are other advantages:

* We will detect malformed annotations sooner in the bootstrap process
* If a given type has multiple indexed subtypes, we will now only parse
its annotations once, instead of once per subtype.
* If we evolve the annotation mapping API (we add annotations), we will
*have* to think about evolving the programmatic API too, since we rely
on it to implement annotation mapping. Thus the risk of having an
outdated programmatic mapping API will be much lower.