HSEARCH-2546 Allow mappers to abort mapping building early if they detect a collected failure
This should avoid snowball effect on failures.

For instance, without this patch, if the document ID was defined on a
type for which we can't find any document ID bridge, we would get two
failures: the first one because the bridge could not be found, and the
second one because no document ID was assigned (which results directly
from the first failure).
With this patch, only the first failure would get reported, because the
mapping building would stop before even trying to validate that a
document ID was defined.