HSEARCH-3239 Move Hibernate ORM DB connection settings to system properties
... so that we can debug more easily. Previously we had to remember to
always re-compile the hibernate-search-util-internal-integrationtest-orm
module when we switch the DB profile, and that's really not something
you can expect from someone that hasn't done it for more than a month.

With the system properties, you just need to recompile the module
containing the tests you want to run.

We could also have moved the properties to one file in each integration
test module, but that would have required to duplicate some configuration,
and I'd rather avoid that.