HSEARCH-3294 Use a more flexible way to configure email notifications in the Jenkinsfile
The recipient providers (developers(), culprits()) are not as reliable
as I though they were: sometimes the list of changes is not available
for a given build for some reason, and in that case they won't work.

Thus, if we want to configure jobs for each forked repository of
Hibernate Search (mine for example), we need to be able to configure
notification recipients for each of those.