HSEARCH-3139 Move Hibernate Search engine bootstrapping to SPI
Engine bootstrapping interfaces reference SPIs, such as BeanResolver,
so we cannot expect users to access them directly. We will expect mapper
implementors to provide access to bootstrapping with APIs of their own.

In cases where we want to allow multiple mappers to be started up in the
same engine, this will still be possible, but each mapper will have to
expose SPI implementations (mainly, mapping keys and mapping initiators)
to allow that.